validando a nivel de servidor todo

+++ app/Http/Controllers/Voyager/ServerController.php
@@ -1053,7 +1053,19 @@ class ServerController extends VoyagerBaseController
         foreach($customers as $customer){
             $server = Server::findorfail($customer->server_id);
             $this->plex->setServerCredentials($server->url, $server->token);
+
+            $plex_data = $this->plex->provider->getAccounts();
+
+            if(!is_array($plex_data)){
+                $redirect = redirect()->back();
+                return $redirect->with([
+                    'message'    => __('Existen problemas en el servidor, por favor verifica que la url del mismo, el puerto y tambien el token sean los correctos!!'),
+                    'alert-type' => 'error',
+                ]);
+            }
+
             $this->plex->provider->removeFriend($customer->invited_id);
+            
             $cus = customer::findorfail($customer->id);
             $cus->delete();
         }
